Issues 351 - Moving TTS Factory to tts __init__.py
