+++ mycroft/tts/__init__.py
@@ -16,9 +16,18 @@
 # along with Mycroft Core.  If not, see <http://www.gnu.org/licenses/>.
 
 
+import logging
+
 import abc
 from os.path import dirname, exists, isdir
 
+from mycroft.configuration import ConfigurationManager
+from mycroft.tts import espeak_tts
+from mycroft.tts import fa_tts
+from mycroft.tts import google_tts
+from mycroft.tts import mary_tts
+from mycroft.tts import mimic_tts
+from mycroft.tts import spdsay_tts
 from mycroft.util.log import getLogger
 
 __author__ = 'jdorleans'
@@ -91,3 +100,44 @@ class TTSValidator(object):
     @abc.abstractmethod
     def get_instance(self):
         pass
+
+
+class TTSFactory(object):
+    @staticmethod
+    def create():
+        """
+        Factory method to create a TTS engine based on configuration.
+
+        The configuration file ``mycroft.conf`` contains a ``tts`` section with
+        the name of a TTS module to be read by this method.
+
+        "tts": {
+            "module": <engine_name>
+        }
+        """
+
+        logging.basicConfig()
+        config = ConfigurationManager.get().get('tts')
+        name = config.get('module')
+        lang = config.get(name).get('lang')
+        voice = config.get(name).get('voice')
+
+        if name == mimic_tts.NAME:
+            tts = mimic_tts.Mimic(lang, voice)
+            mimic_tts.MimicValidator().validate(tts)
+        elif name == google_tts.NAME:
+            tts = google_tts.GoogleTTS(lang, voice)
+            google_tts.GoogleTTSValidator().validate(tts)
+        elif name == mary_tts.NAME:
+            tts = mary_tts.MaryTTS(lang, voice, config[name + '.url'])
+            mary_tts.MaryTTSValidator().validate(tts)
+        elif name == fa_tts.NAME:
+            tts = fa_tts.FATTS(lang, voice, config[name + '.url'])
+            fa_tts.FATTSValidator().validate(tts)
+        elif name == espeak_tts.NAME:
+            tts = espeak_tts.ESpeak(lang, voice)
+            espeak_tts.ESpeakValidator().validate(tts)
+        else:
+            tts = spdsay_tts.SpdSay(lang, voice)
+            spdsay_tts.SpdSayValidator().validate(tts)
+        return tts
